+++ tests/Application/Contracts/Bitrix24Partners/Entity/Bitrix24PartnerInterfaceTest.php
@@ -15,11 +15,9 @@ namespace Bitrix24\SDK\Tests\Application\Contracts\Bitrix24Partners\Entity;
 
 use Bitrix24\SDK\Application\Contracts\Bitrix24Partners\Entity\Bitrix24PartnerInterface;
 use Bitrix24\SDK\Application\Contracts\Bitrix24Partners\Entity\Bitrix24PartnerStatus;
-use Bitrix24\SDK\Application\Contracts\Bitrix24Partners\Events\Bitrix24PartnerExternalIdChangedEvent;
 use Bitrix24\SDK\Application\Contracts\Events\AggregateRootEventsEmitterInterface;
 use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
 use Bitrix24\SDK\Tests\Builders\DemoDataGenerator;
-use Carbon\CarbonImmutable;
 use Generator;
 use libphonenumber\NumberParseException;
 use libphonenumber\PhoneNumber;
@@ -29,15 +27,12 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\Attributes\TestDox;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Uid\Uuid;
-use Throwable;
 
 #[CoversClass(Bitrix24PartnerInterface::class)]
 abstract class Bitrix24PartnerInterfaceTest extends TestCase
 {
     abstract protected function createBitrix24PartnerImplementation(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -45,7 +40,8 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?PhoneNumber          $phoneNumber,
         ?string               $email,
         ?string               $openLineId,
-        ?string               $externalId
+        ?string               $externalId,
+        ?string               $logoUrl = null
     ): Bitrix24PartnerInterface|AggregateRootEventsEmitterInterface;
 
     #[Test]
@@ -53,8 +49,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getId method')]
     final public function testGetId(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -63,33 +57,11 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($uuid, $bitrix24Partner->getId());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test getCreatedAt method')]
-    final public function testGetCreatedAt(
-        Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        Bitrix24PartnerStatus $bitrix24PartnerStatus,
-        string                $title,
-        int                   $bitrix24PartnerNumber,
-        ?string               $site,
-        ?PhoneNumber          $phoneNumber,
-        ?string               $email,
-        ?string               $openLineId,
-        ?string               $externalId,
-        string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
-        $this->assertEquals($createdAt, $bitrix24Partner->getCreatedAt());
     }
 
     /**
@@ -100,8 +72,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getUpdatedAt method')]
     final public function testGetUpdatedAt(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -110,21 +80,19 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
-        $bitrix24Partner->setTitle('new title');
-        $this->assertNotEquals($updatedAt, $bitrix24Partner->getUpdatedAt());
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
+        $bitrix24Partner->changeTitle('new title');
+        $this->assertFalse($bitrix24Partner->getUpdatedAt()->equalTo($bitrix24Partner->getCreatedAt()));
     }
 
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setExternalId method with empty string')]
-    final public function testSetExternalId(
+    #[TestDox('test changeExternalId method with empty string')]
+    final public function testChangeExternalId(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -133,13 +101,13 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
 
         $newExternalId = Uuid::v7()->toRfc4122();
-        $bitrix24Partner->setExternalId($newExternalId);
+        $bitrix24Partner->changeExternalId($newExternalId);
         $this->assertEquals($newExternalId, $bitrix24Partner->getExternalId());
     }
 
@@ -148,11 +116,9 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
      */
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setExternalId method with null')]
-    final public function testSetExternalIdWithNull(
+    #[TestDox('test changeExternalId method with null')]
+    final public function testChangeExternalIdWithNull(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -161,21 +127,19 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
-        $bitrix24Partner->setExternalId(null);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
+        $bitrix24Partner->changeExternalId(null);
         $this->assertNull($bitrix24Partner->getExternalId());
     }
 
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setExternalId method')]
-    final public function testSetExternalIdWithEmptyString(
+    #[TestDox('test changeExternalId method')]
+    final public function testChangeExternalIdWithEmptyString(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -184,13 +148,13 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->expectException(InvalidArgumentException::class);
         /** @phpstan-ignore-next-line */
-        $bitrix24Partner->setExternalId('');
+        $bitrix24Partner->changeExternalId('');
     }
 
     #[Test]
@@ -198,8 +162,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getStatus method')]
     final public function testGetStatus(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -208,10 +170,10 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($bitrix24PartnerStatus, $bitrix24Partner->getStatus());
     }
 
@@ -223,8 +185,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test markAsActive method')]
     final public function testMarkAsActive(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -233,10 +193,10 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
 
         $blockComment = 'block partner';
         $bitrix24Partner->markAsBlocked($blockComment);
@@ -256,8 +216,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test markAsBlocked method')]
     final public function testMarkAsBlocked(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -266,10 +224,10 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
 
         $blockComment = 'block partner';
         $bitrix24Partner->markAsBlocked($blockComment);
@@ -285,8 +243,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test markAsBlocked method')]
     final public function testMarkAsDeleted(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -295,10 +251,10 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
 
         $comment = 'delete partner';
         $bitrix24Partner->markAsDeleted($comment);
@@ -314,8 +270,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getTitle method')]
     final public function testGetTitle(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -324,20 +278,18 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($title, $bitrix24Partner->getTitle());
     }
 
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setTitle method')]
-    final public function testSetTitle(
+    #[TestDox('test changeTitle method')]
+    final public function testChangeTitle(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -346,18 +298,18 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $newTitle = 'new title';
-        $bitrix24Partner->setTitle($newTitle);
+        $bitrix24Partner->changeTitle($newTitle);
         $this->assertEquals($newTitle, $bitrix24Partner->getTitle());
 
         $newTitle = '';
         $this->expectException(InvalidArgumentException::class);
         /** @phpstan-ignore-next-line */
-        $bitrix24Partner->setTitle($newTitle);
+        $bitrix24Partner->changeTitle($newTitle);
     }
 
     #[Test]
@@ -365,8 +317,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getSite method')]
     final public function testGetSite(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -375,10 +325,10 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($site, $bitrix24Partner->getSite());
     }
 
@@ -387,11 +337,9 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
      */
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setSite method')]
-    final public function testSetSite(
+    #[TestDox('test changeSite method')]
+    final public function testChangeSite(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -400,20 +348,20 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $newSite = 'https://new-partner-site.com';
-        $bitrix24Partner->setSite($newSite);
+        $bitrix24Partner->changeSite($newSite);
         $this->assertEquals($newSite, $bitrix24Partner->getSite());
 
-        $bitrix24Partner->setSite(null);
+        $bitrix24Partner->changeSite(null);
         $this->assertNull($bitrix24Partner->getSite());
 
         $this->expectException(InvalidArgumentException::class);
         /** @phpstan-ignore-next-line */
-        $bitrix24Partner->setSite('');
+        $bitrix24Partner->changeSite('');
     }
 
     #[Test]
@@ -421,8 +369,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getPhone method')]
     final public function testGetPhone(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -431,20 +377,18 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($phoneNumber, $bitrix24Partner->getPhone());
     }
 
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setPhone method')]
-    final public function testSetPhone(
+    #[TestDox('test changePhone method')]
+    final public function testChangePhone(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -453,15 +397,15 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $newPhone = DemoDataGenerator::getMobilePhone();
-        $bitrix24Partner->setPhone($newPhone);
+        $bitrix24Partner->changePhone($newPhone);
         $this->assertEquals($newPhone, $bitrix24Partner->getPhone());
 
-        $bitrix24Partner->setPhone(null);
+        $bitrix24Partner->changePhone(null);
         $this->assertNull($bitrix24Partner->getPhone());
     }
 
@@ -470,8 +414,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getEmail method')]
     final public function testGetEmail(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -480,20 +422,18 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($email, $bitrix24Partner->getEmail());
     }
 
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setEmail method')]
-    final public function testSetEmail(
+    #[TestDox('test changeEmail method')]
+    final public function testChangeEmail(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -502,30 +442,28 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
 
         $newEmail = DemoDataGenerator::getEmail();
-        $bitrix24Partner->setEmail($newEmail);
+        $bitrix24Partner->changeEmail($newEmail);
         $this->assertEquals($newEmail, $bitrix24Partner->getEmail());
 
-        $bitrix24Partner->setEmail(null);
+        $bitrix24Partner->changeEmail(null);
         $this->assertNull($bitrix24Partner->getEmail());
 
         $this->expectException(InvalidArgumentException::class);
         /** @phpstan-ignore-next-line */
-        $bitrix24Partner->setEmail('');
+        $bitrix24Partner->changeEmail('');
     }
 
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setEmail method with invalid email')]
-    final public function testSetEmailWithInvalidEmail(
+    #[TestDox('test changeEmail method with invalid email')]
+    final public function testChangeEmailWithInvalidEmail(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -534,14 +472,14 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
 
         $newEmail = '@partner.com';
         $this->expectException(InvalidArgumentException::class);
-        $bitrix24Partner->setEmail($newEmail);
+        $bitrix24Partner->changeEmail($newEmail);
     }
 
     #[Test]
@@ -549,8 +487,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getBitrix24PartnerNumber')]
     final public function testGetBitrix24PartnerNumber(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -559,10 +495,10 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($bitrix24PartnerNumber, $bitrix24Partner->getBitrix24PartnerNumber());
     }
 
@@ -571,8 +507,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     #[TestDox('test getOpenLineId')]
     final public function testGetOpenLineId(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -581,20 +515,18 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $this->assertEquals($openLineId, $bitrix24Partner->getOpenLineId());
     }
 
     #[Test]
     #[DataProvider('bitrix24PartnerDataProvider')]
-    #[TestDox('test setOpenLineId')]
-    final public function testSetOpenLineId(
+    #[TestDox('test changeOpenLineId')]
+    final public function testChangeOpenLineId(
         Uuid                  $uuid,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
         Bitrix24PartnerStatus $bitrix24PartnerStatus,
         string                $title,
         int                   $bitrix24PartnerNumber,
@@ -603,20 +535,73 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
         ?string               $email,
         ?string               $openLineId,
         ?string               $externalId,
+        ?string               $logoUrl,
         string                $comment
-    ): void
-    {
-        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $createdAt, $updatedAt, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId);
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
         $newOpenLineId = Uuid::v7()->toRfc4122();
-        $bitrix24Partner->setOpenLineId($newOpenLineId);
+        $bitrix24Partner->changeOpenLineId($newOpenLineId);
         $this->assertEquals($newOpenLineId, $bitrix24Partner->getOpenLineId());
 
-        $bitrix24Partner->setOpenLineId(null);
+        $bitrix24Partner->changeOpenLineId(null);
         $this->assertNull($bitrix24Partner->getOpenLineId());
 
         $this->expectException(InvalidArgumentException::class);
         /** @phpstan-ignore-next-line */
-        $bitrix24Partner->setOpenLineId('');
+        $bitrix24Partner->changeOpenLineId('');
+    }
+
+    #[Test]
+    #[DataProvider('bitrix24PartnerDataProvider')]
+    #[TestDox('test getLogoUrl method')]
+    final public function testGetLogoUrl(
+        Uuid                  $uuid,
+        Bitrix24PartnerStatus $bitrix24PartnerStatus,
+        string                $title,
+        int                   $bitrix24PartnerNumber,
+        ?string               $site,
+        ?PhoneNumber          $phoneNumber,
+        ?string               $email,
+        ?string               $openLineId,
+        ?string               $externalId,
+        ?string               $logoUrl,
+        string                $comment
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
+        $this->assertEquals($logoUrl, $bitrix24Partner->getLogoUrl());
+    }
+
+    /**
+     * @throws InvalidArgumentException
+     */
+    #[Test]
+    #[DataProvider('bitrix24PartnerDataProvider')]
+    #[TestDox('test changeLogoUrl method')]
+    final public function testChangeLogoUrl(
+        Uuid                  $uuid,
+        Bitrix24PartnerStatus $bitrix24PartnerStatus,
+        string                $title,
+        int                   $bitrix24PartnerNumber,
+        ?string               $site,
+        ?PhoneNumber          $phoneNumber,
+        ?string               $email,
+        ?string               $openLineId,
+        ?string               $externalId,
+        ?string               $logoUrl,
+        string                $comment
+    ): void {
+        $bitrix24Partner = $this->createBitrix24PartnerImplementation($uuid, $bitrix24PartnerStatus, $title, $bitrix24PartnerNumber, $site, $phoneNumber, $email, $openLineId, $externalId, $logoUrl);
+
+        $newLogoUrl = 'https://new-partner-logo.example.com/logo.png';
+        $bitrix24Partner->changeLogoUrl($newLogoUrl);
+        $this->assertEquals($newLogoUrl, $bitrix24Partner->getLogoUrl());
+
+        $bitrix24Partner->changeLogoUrl(null);
+        $this->assertNull($bitrix24Partner->getLogoUrl());
+
+        $this->expectException(InvalidArgumentException::class);
+        /** @phpstan-ignore-next-line */
+        $bitrix24Partner->changeLogoUrl('');
     }
 
     /**
@@ -627,8 +612,6 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
     {
         yield 'partner-status-active-all-fields' => [
             Uuid::v7(), //id
-            CarbonImmutable::now(), // createdAt
-            CarbonImmutable::now(), // updatedAt
             Bitrix24PartnerStatus::active,
             'Bitrix24 Partner LLC', // title
             12345, // bitrix24 partner number, optional
@@ -637,6 +620,7 @@ abstract class Bitrix24PartnerInterfaceTest extends TestCase
             DemoDataGenerator::getEmail(), // email, optional
             'open-line-id', // open line id, optional
             Uuid::v7()->toRfc4122(), // externalId, optional
+            'https://partner-logo.example.com/logo.png', // logoUrl, optional
             'comment', // comment, optional
         ];
     }